Expose render and renderHtml methods
